+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `override.piraeus.io` annotations to PVs to override access context.
+
 ## [1.2.0] - 2025-06-27
 
 ### Added